Relocate sortTypedPhoneNumbers and validateTypedPhones to shared modules
Moves sortTypedPhoneNumbers (and its compareTypedPhoneNumbers helper)
from common/cams/trustees to common/cams/contact, where TypedPhoneNumber
and its sibling types already live. Moves validateTypedPhones from
trustees/forms/trusteeForms.utils to a new phoneEntryList.utils module
colocated with PhoneEntryList.

Both trustees.ts and trusteeForms.utils.ts now re-export from the new
canonical homes, so existing callers are unaffected. All cross-feature
imports in bankruptcy-software and shared lib components are updated to
reference contact.ts and phoneEntryList.utils directly.

No behavior change — pure relocation.

Jira ticket: CAMS-825
