+++ common/src/cams/trustees.ts
@@ -1,14 +1,7 @@
 import { Auditable } from './auditable';
 import { Identifiable } from './document';
 import { LegacyAddress } from './parties';
-import {
-  Address,
-  ContactInformation,
-  Person,
-  PhoneNumber,
-  TypedPhoneNumber,
-  PHONE_TYPES,
-} from './contact';
+import { Address, ContactInformation, Person, PhoneNumber, TypedPhoneNumber } from './contact';
 import { CamsUserReference } from './users';
 import { OversightRoleType } from './roles';
 import { NullableOptionalFields } from '../api/common';
@@ -18,24 +11,15 @@ import { TrusteeUpcomingKeyDatesHistory } from './trustee-upcoming-key-dates';
 import type { TrusteeAppointment } from './trustee-appointments';
 
 export type { PhoneType, TypedPhoneNumber } from './contact';
-export { PHONE_TYPES, PHONE_TYPE_LABELS, MAX_PHONE_NUMBERS } from './contact';
+export {
+  PHONE_TYPES,
+  PHONE_TYPE_LABELS,
+  MAX_PHONE_NUMBERS,
+  sortTypedPhoneNumbers,
+} from './contact';
 export type TrusteeContact = Omit<Partial<ContactInformation>, 'phone'> & {
   phones?: TypedPhoneNumber[];
 };
-
-function compareTypedPhoneNumbers(a: TypedPhoneNumber, b: TypedPhoneNumber): number {
-  const typeCompare = PHONE_TYPES.indexOf(a.type) - PHONE_TYPES.indexOf(b.type);
-  if (typeCompare !== 0) return typeCompare;
-
-  const numberCompare = a.number.localeCompare(b.number);
-  if (numberCompare !== 0) return numberCompare;
-
-  return (a.extension ?? '').localeCompare(b.extension ?? '');
-}
-
-export function sortTypedPhoneNumbers(phones: TypedPhoneNumber[]): TypedPhoneNumber[] {
-  return [...phones].sort(compareTypedPhoneNumbers);
-}
 
 export function sortTrusteePhoneNumbers(trustee: Trustee): Trustee {
   return {
